+++ example/main.dart
@@ -1,0 +1,22 @@
+// Copyright 2021 Andrious Solutions Ltd. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+
+import 'src/employee_list.dart';
+
+void main() => runApp(const MyApp());
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+  @override
+  Widget build(BuildContext context) => MaterialApp(
+        title: 'DBUtils Demo app',
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
+        ),
+        home: const EmployeeList(),
+      );
+}